requirements.txt
tensorflow==2.15.0
opencv-python==4.8.1.78
numpy==1.24.3
flask==3.0.0
pillow==10.1.0
facenet-pytorch==2.5.3
mtcnn==0.1.1
gunicorn==21.2.0 